Moved can_unsubscribe_topic_key
